AAAAAAA

+++ linha.c
@@ -6,11 +6,11 @@
 
 typedef struct {
     int id;
-    int x1, y1, x2, y2;
+    double x1, y1, x2, y2;
     char* cor;
 }linha;
 
-LINHA cria_linha(int i, int x_1, int x_2, int y_1, int y_2, char* corL){
+LINHA cria_linha(int i, double x_1, double x_2, double y_1, double y_2, char* corL){
     linha* l = malloc(sizeof(linha));
 
     if(l==NULL){
@@ -40,24 +40,24 @@ int get_id_linha(LINHA L){
     return l->id;
 }
 
-int get_codx1_linha(LINHA L){  
+double get_codx1_linha(LINHA L){  
     linha* l = ((linha*)L);
     return l->x1;
 }
 
 
-int get_codx2_linha(LINHA L){  
+double get_codx2_linha(LINHA L){  
     linha* l = ((linha*)L);
     return l->x2;
 }
 
-int get_cody1_linha(LINHA L){  
+double get_cody1_linha(LINHA L){  
     linha* l = ((linha*)L);
     return l->y1;
 }
 
 
-int get_cody2_linha(LINHA L){  
+double get_cody2_linha(LINHA L){  
     linha* l = ((linha*)L);
     return l->y2;
 }
@@ -70,24 +70,24 @@ char* get_cor_linha(LINHA L){
 
 
 
-void set_codx1_linha(LINHA L, int x){  
+void set_codx1_linha(LINHA L, double x){  
     linha* l = ((linha*)L);
     l->x1 = x;
 }
 
 
-void set_codx2_linha(LINHA L, int x){  
+void set_codx2_linha(LINHA L, double x){  
     linha* l = ((linha*)L);
     l->x2 = x;
 }
 
-void set_cody1_linha(LINHA L, int y){  
+void set_cody1_linha(LINHA L, double y){  
     linha* l = ((linha*)L);
     l->y1 = y;
 }
 
 
-void set_cody2_linha(LINHA L, int y){  
+void set_cody2_linha(LINHA L, double y){  
     linha* l = ((linha*)L);
     l->y2 = y;
 }
