+++ texto.h
@@ -7,14 +7,14 @@ typedef void *Estilo;
 
 Estilo cria_estilo(char* Family, char* Weight, char* Size );
 
-TEXTO cria_texto( int i, int codX, int codY, char *cor_b, char *cor_p, char *a, char *TXTO);
+TEXTO cria_texto( int i, double codX, double codY, char *cor_b, char *cor_p, char *a, char *TXTO);
 
 
 int get_id_txt(TEXTO t);
 
-int get_codx_txt(TEXTO t);
+double get_codx_txt(TEXTO t);
 
-int get_cody_txt(TEXTO t);
+double get_cody_txt(TEXTO t);
 
 char* get_corB_txt(TEXTO t);
 
@@ -31,9 +31,9 @@ char* getWeight(Estilo e);
 char* getSize(Estilo e);
 
 
-void set_codx_txt(TEXTO t, int Cx);
+void set_codx_txt(TEXTO t, double Cx);
 
-void set_cody_txt(TEXTO t, int Cy);
+void set_cody_txt(TEXTO t, double Cy);
 
 void set_corB_txt(TEXTO t, char* cor_b);
